カメラ機能微調整
他のシーンでエラーが起こらないように
