+++ Assets/Scripts/CameraManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d73de097b516e64b9c72876fb66db7b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
