+++ UserSettings/EditorUserSettings.asset
@@ -21,10 +21,10 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c6b0203122c23382423032e2c1336acf53a31f6fe
       flags: 0
     RecentlyUsedScenePath-5:
-      value: 22424703114646680e0b0227036c7c1f191c11242b1b2535232c5326ece92021
+      value: 22424703114646680e0b0227036c7c1817051929382d34032e2c1336acf53a31f6fe
       flags: 0
     RecentlyUsedScenePath-6:
-      value: 22424703114646680e0b0227036c7c1817051929382d34032e2c1336acf53a31f6fe
+      value: 22424703114646680e0b0227036c7c1f191c11242b1b2535232c5326ece92021
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
